requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
Django==5.2.4
djangorestframework==3.16.0
flake8==7.3.0
inflection==0.5.1
isort==6.0.1
mccabe==0.7.0
packaging==25.0
pycodestyle==2.14.0
pyflakes==3.4.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
